Made seed dir in tasks and renamed local seed file
